% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-ggplot.R
\name{make_mtp_tbl}
\alias{make_mtp_tbl}
\title{The idea here is to have a function that generates the following elements to represent a microplate:
\enumerate{
\item footprint: this is a rounded rect that is the outer border of the microwell plate, it's footprint.
\item notched border: this is the border that has notches in it.
\item wells: this is either circles or rounded rect elements (one per well)
\item labels: row and column labels for the microwell plate, text elements.
}}
\usage{
make_mtp_tbl(nrow, ncol, wells_as = "rects")
}
\arguments{
\item{wells_as}{<chr> "rects" or "circles"}

\item{n_wells}{<int> number of wells}
}
\description{
The idea here is to have a function that generates the following elements to represent a microplate:
\enumerate{
\item footprint: this is a rounded rect that is the outer border of the microwell plate, it's footprint.
\item notched border: this is the border that has notches in it.
\item wells: this is either circles or rounded rect elements (one per well)
\item labels: row and column labels for the microwell plate, text elements.
}
}
\examples{
  make_microwell_plate_tbl(384, wells_as = "rects")
}
