% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-ggplot.R
\name{mtp_footprint}
\alias{mtp_footprint}
\title{Rounded rectangle of microtitreplate footprint from width and height}
\usage{
mtp_footprint(w, h, r = 3)
}
\arguments{
\item{w}{width in mm}

\item{h}{height in mm}

\item{r}{radius of circle used to round corners in mm}
}
\description{
Rounded rectangle of microtitreplate footprint from width and height
}
\examples{
    mtp_footprint(127.76, 85.48)
}
