% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-ggplot.R
\name{mtp_ggplot}
\alias{mtp_ggplot}
\title{plot a microtitre plate ggplot style}
\usage{
mtp_ggplot(tbl, well_var = "well", well_fill_var = NULL,
  well_alpha_var = NULL, draw_lineplots = FALSE)
}
\arguments{
\item{tbl}{table of data to operate on, one row per well.}

\item{well_var}{name of the variable in tbl that has well identifiers (e.g. 'A01')}

\item{well_fill_var}{name of variable in tbl to map to well fill}

\item{well_alpha_var}{name of variable in tbl to map to well transparency (alpha)}
}
\description{
plot a microtitre plate ggplot style
}
\examples{
    tbl <- read_tsv('test-plate.tsv') \%>\%  nest(runtime, measure)
    tbl <- tbl \%>\% mutate(cond = case_when(
                                           is.na(drug_name) ~ 'ref',
                                           drug_name == 'Blank' ~ 'blank',
                                           TRUE ~ 'trt'))
    ggplate(tbl, well, )
}
