% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-ggplot.R
\name{mtp_notched_border}
\alias{mtp_notched_border}
\title{Path for rectangle with notched border of microtitre plate.}
\usage{
mtp_notched_border(w, h)
}
\arguments{
\item{w}{width in cm}

\item{h}{height in cm}
}
\description{
Path for rectangle with notched border of microtitre plate.
}
\examples{
    mtp_notched_border(127.76, 85.48)
}
