% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmuter.R
\name{mtp_wells_tbl}
\alias{mtp_wells_tbl}
\title{return a one-row-per-well tibble of a microtitre plate with n rows and n cols.}
\usage{
mtp_wells_tbl(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows in microtitre plate}

\item{ncol}{number of columns in microtitre plate}
}
\description{
return a one-row-per-well tibble of a microtitre plate with n rows and n cols.
}
\examples{
    mtp_wells_tbl(6, 8)
    mtp_wells_tbl(8, 12)
    mtp_wells_tbl(16, 24)
    mtp_wells_tbl(32, 48)
    mtp_wells_tbl(64, 96)
}
