% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-do.R
\name{rescale_to_wells}
\alias{rescale_to_wells}
\title{rescale x,y to the coordinate system of the destination well
for plotting. to_x, to_y, width, height, are the attributes of
the destination well to scale against
by default this scales from the ranges of x and y to the
range of the rectangular well. Need to come up with a way to
offer well-rescaling, plate-rescaling, experiment-rescaling.}
\usage{
rescale_to_wells(x, y, to_x, to_y, width, height)
}
\description{
rescale x,y to the coordinate system of the destination well
for plotting. to_x, to_y, width, height, are the attributes of
the destination well to scale against
by default this scales from the ranges of x and y to the
range of the rectangular well. Need to come up with a way to
offer well-rescaling, plate-rescaling, experiment-rescaling.
}
