% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandscale.R
\name{show_bandscale}
\alias{show_bandscale}
\title{show_bandscale(bandscale_01(3, 0.05, 1))
show_bandscale(bandscale_01(12, 0.1, 0.5))
show_bandscale(bandscale_01(24, 0.1, 0.5))
show_bandscale(bandscale_01(96, 0.1, 0.5))}
\usage{
show_bandscale(l)
}
\description{
show_bandscale(bandscale_01(3, 0.05, 1))
show_bandscale(bandscale_01(12, 0.1, 0.5))
show_bandscale(bandscale_01(24, 0.1, 0.5))
show_bandscale(bandscale_01(96, 0.1, 0.5))
}
