% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtp-ggplot.R
\name{mtp_wells}
\alias{mtp_wells}
\title{plot elements for wells as rects or circles based on plate params.
One row per well output.}
\usage{
mtp_wells(nrow, ncol, cx_shift, cy_shift, well_c2c, well_radius, wells_as)
}
\description{
plot elements for wells as rects or circles based on plate params.
One row per well output.
}
